requirements.txt
#torch
torch-geometric
torch-scatter
torch-sparse
torch-cluster
selenium
pillow
gym
networkx
beautifulsoup4
bpemb
git+https://github.com/openai/baselines.git@master#egg=baselines
linkGrabber
tensorboardX
arsenic
pytest
pytest-asyncio
git+https://github.com/ikostrikov/pytorch-a2c-ppo-acktr-gail.git@master#egg=pytorch-a2c-ppo-acktr-gail
git+https://github.com/zbyte64/miniwob-plusplus.git@89814f4#egg=miniwob
